Fix a bug when a media is changded to IFM_AUTO.
 Fix a bug that ifconfig ifN media auto doesn't change the setting when
the previous media setting used autonego. When the mii_phy_setmedia() function
is called to chage the media to IFM_AUTO, the BMCR_AUTOEN bit is used to
check if the previos setting was IFM_AUTO. It's not correct. IFM_1000_T
also uses autonego. So if a previous setting is IFM_1000_T and the next
setting is IFM_AUTO, mii_phy_auto() is not called if neither MIIF_FORCEANEG
nor MIIF_DOPAUSE are set. As a result, after changing IFM_AUTO, neither
10Mbps nor 100Mbps are not advertised.

 Note that almost all drivers uses MIIF_DOPAUSE flags.